service/requirements.txt
fastapi
uvicorn[standard]
asyncpg
aiohttp
jinja2
python-dotenv
plotly
pandas
